pyproject.toml
[tool.poetry]
name = "instaworks"
version = "0.1.0"
description = ""
authors = ["adrian <adrian@bbfmedia.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.3"
djangorestframework = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
